+++ packages/connectivity/src/scp-cf/destination/index.ts
@@ -15,5 +15,6 @@ export * from './get-subscriber-token';
 export * from './get-provider-token';
 export * from './http-proxy-util';
 export * from './ias-types';
+export * from './build-ias-destination';
 export * from './service-binding-to-destination';
 export * from './register-destination-cache';
